LayoutTests/http/wpt/web-animations/interfaces/Animation/startTime-expected.txt


PASS startTime of a newly created (idle) animation is unresolved 
FAIL startTime of a play-pending animation is unresolved assert_equals: startTime is unresolved expected (object) null but got (number) 48.300000000000004
PASS startTime of a pause-pending animation is unresolved 
FAIL startTime of a play-pending animation created using Element.animate shortcut is unresolved assert_equals: startTime is unresolved expected (object) null but got (number) 48.300000000000004
PASS startTime is resolved when running 
PASS startTime and currentTime are unresolved when animation is cancelled 
